station_manager/requirements.txt
pandas==1.1.4
requests==2.25.0
google-cloud-bigquery==2.4.0
google-cloud-storage==1.32.0